+++ packages/ws-worker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ws-worker
 
+## 1.1.0
+
+Allow runs to use multiple versions of the same adaptor
+
+### Patch Changes
+
+- 58e0d11: Move version log to workflow start
+- Updated dependencies [4f5f1dd]
+- Updated dependencies [58e0d11]
+  - @openfn/engine-multi@1.1.0
+  - @openfn/runtime@1.1.0
+
 ## 1.0.0
 
 ### Major Changes